Fixed auth guard permission handling; added 404 route
