+++ src/components/gates/requireUnauth.tsx
@@ -11,8 +11,7 @@ function RequireUnauth({children}: { children: JSX.Element}) {
     const {user} = useAuth();
     const location = useLocation() as LocationState;
 
-    console.log("User: ", user)
-
+    console.log('requireUnauth')
     if (user.loading) return <Loading open={true} />;
     if (user.isAuthenticated) return <Navigate to={location.state?.from || '/'} replace />;
     else return children;
